refactor: move some components to more appropriate location
