+++ src/components/Header/index.tsx
@@ -10,8 +10,8 @@ import { LanguageDropdown } from './LanguageComp'
 import BlockchainComp from './BlockchainComp'
 import { useElementSize, useIsMobile } from '../../hooks'
 import styles from './index.module.scss'
-import MaintainAlert from '../MaintainAlert'
-import Sheet from '../Sheet'
+import MaintainAlert from './MaintainAlert'
+import Sheet from './Sheet'
 import { createGlobalState, useGlobalState } from '../../utils/state'
 import MobileMenu from './MobileMenu'
