+++ src/components/QueryResult/Error/index.tsx
@@ -1,6 +1,6 @@
 import PCDataNotFoundImage from './pc_data_not_found.png'
 import MobileDataNotFoundImage from './mobile_data_not_found.png'
-import { useIsMobile } from '../../hooks'
+import { useIsMobile } from '../../../hooks'
 import { ErrorPanel } from './styled'
 
 export default () => {
