+++ src/components/QueryResult/Error/styled.tsx
