+++ src/components/QueryResult/index.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from 'react'
 import { DefinedUseQueryResult } from '@tanstack/react-query'
 import { LOADING_WAITING_TIME } from '../../constants/common'
-import Error from '../Error'
+import Error from './Error'
 import Loading from '../Loading'
 import { useDelayLoading } from '../../hooks'
