+++ src/pages/Home/Banner/halving_banner_success_mobile.png
