+++ src/pages/Home/Banner/halving_success_ani.gif
