+++ src/pages/Home/Banner/index.module.scss
