+++ src/pages/Home/index.tsx
@@ -23,8 +23,8 @@ import LatestTransactionsIcon from './latest_transactions.png'
 import { BlockCardItem, TransactionCardItem } from './TableCard'
 import Loading from '../../components/Loading/SmallLoading'
 import { useElementIntersecting, useIsLGScreen, useIsMobile, useSingleHalving } from '../../hooks'
-import Banner from '../../components/Banner'
-import { HalvingBanner } from '../../components/Banner/HalvingBanner'
+import Banner from './Banner'
+import { HalvingBanner } from './Banner/HalvingBanner'
 import Search from '../../components/Search'
 import AverageBlockTimeChart from './AverageBlockTimeChart'
 import HashRateChart from './HashRateChart'
