+++ src/pages/NftCollectionInfo/NftCollectionInventory/index.tsx
@@ -3,13 +3,13 @@ import { useQuery } from '@tanstack/react-query'
 import { Base64 } from 'js-base64'
 import { hexToBytes } from '@nervosnetwork/ckb-sdk-utils'
 import { useTranslation } from 'react-i18next'
-import { parseSporeCellData } from '../../utils/spore'
-import { ReactComponent as Cover } from '../../assets/nft_cover.svg'
+import { parseSporeCellData } from '../../../utils/spore'
+import { ReactComponent as Cover } from '../../../assets/nft_cover.svg'
 import styles from './styles.module.scss'
-import { getPrimaryColor } from '../../constants/common'
-import { explorerService } from '../../services/ExplorerService'
-import { handleNftImgError, patchMibaoImg } from '../../utils/util'
-import type { NFTItem } from '../../services/ExplorerService/fetcher'
+import { getPrimaryColor } from '../../../constants/common'
+import { explorerService } from '../../../services/ExplorerService'
+import { handleNftImgError, patchMibaoImg } from '../../../utils/util'
+import type { NFTItem } from '../../../services/ExplorerService/fetcher'
 
 const primaryColor = getPrimaryColor()
