+++ src/pages/NftCollectionInfo/NftCollectionInventory/styles.module.scss
