+++ src/pages/NftCollectionInfo/NftCollectionOverview/index.tsx
@@ -2,10 +2,10 @@ import { Link } from 'react-router-dom'
 import { useQuery } from '@tanstack/react-query'
 import { Tooltip } from 'antd'
 import { useTranslation } from 'react-i18next'
-import { explorerService } from '../../services/ExplorerService'
+import { explorerService } from '../../../services/ExplorerService'
 import styles from './styles.module.scss'
-import { handleNftImgError, patchMibaoImg } from '../../utils/util'
-import { getPrimaryColor } from '../../constants/common'
+import { handleNftImgError, patchMibaoImg } from '../../../utils/util'
+import { getPrimaryColor } from '../../../constants/common'
 
 const primaryColor = getPrimaryColor()
