+++ src/pages/NftCollectionInfo/NftCollectionOverview/styles.module.scss
