+++ src/pages/NftCollectionInfo/NftCollectionTransfers/index.tsx
@@ -5,16 +5,16 @@ import { Tooltip } from 'antd'
 import { Base64 } from 'js-base64'
 import { hexToBytes } from '@nervosnetwork/ckb-sdk-utils'
 import { useTranslation } from 'react-i18next'
-import { parseSporeCellData } from '../../utils/spore'
+import { parseSporeCellData } from '../../../utils/spore'
 // TODO: Refactor is needed. Should not directly import anything from the descendants of ExplorerService.
-import type { TransferListRes, TransferRes } from '../../services/ExplorerService/fetcher'
+import type { TransferListRes, TransferRes } from '../../../services/ExplorerService/fetcher'
 import styles from './styles.module.scss'
-import { getPrimaryColor } from '../../constants/common'
-import { handleNftImgError, patchMibaoImg } from '../../utils/util'
-import { explorerService } from '../../services/ExplorerService'
-import { dayjs } from '../../utils/date'
-import { useParsedDate, useTimestamp } from '../../hooks'
-import { useCurrentLanguage } from '../../utils/i18n'
+import { getPrimaryColor } from '../../../constants/common'
+import { handleNftImgError, patchMibaoImg } from '../../../utils/util'
+import { explorerService } from '../../../services/ExplorerService'
+import { dayjs } from '../../../utils/date'
+import { useParsedDate, useTimestamp } from '../../../hooks'
+import { useCurrentLanguage } from '../../../utils/i18n'
 
 const primaryColor = getPrimaryColor()
