+++ src/pages/NftCollectionInfo/NftHolderList/styles.module.scss
