+++ src/pages/NftCollectionInfo/index.tsx
@@ -3,9 +3,9 @@ import { useQuery } from '@tanstack/react-query'
 import { Popover } from 'antd'
 import { TFunction, useTranslation } from 'react-i18next'
 import Content from '../../components/Content'
-import NftCollectionOverview from '../../components/NftCollectionOverview'
-import NftCollectionTransfers from '../../components/NftCollectionTransfers'
-import NftCollectionInventory from '../../components/NftCollectionInventory'
+import NftCollectionOverview from './NftCollectionOverview'
+import NftCollectionTransfers from './NftCollectionTransfers'
+import NftCollectionInventory from './NftCollectionInventory'
 import Filter from '../../components/Search/Filter'
 import { ReactComponent as FilterIcon } from '../../assets/filter_icon.svg'
 import { ReactComponent as SelectedCheckIcon } from '../../assets/selected_check_icon.svg'
@@ -14,7 +14,7 @@ import { useSearchParams, useIsMobile } from '../../hooks'
 import styles from './styles.module.scss'
 import { CsvExport } from '../../components/CsvExport'
 import PaginationWithRear from '../../components/PaginationWithRear'
-import NftHolderList from '../../components/NftHolderList'
+import NftHolderList from './NftHolderList'
 import Pagination from '../../components/Pagination'
 
 const tabs = ['transfers', 'holders', 'inventory']
