+++ src/pages/NftInfo/NftItemTransfers/styles.module.scss
